+++ cmd/osv-scanner/main.go
@@ -91,23 +91,22 @@ NAME:
 	{{.Name}} - {{.Usage}}
 
 USAGE:
-	{{.HelpName}} {{if .VisibleFlags}}[global options]{{end}}{{if .Commands}} command [command options]{{end}}
+	{{.Name}} {{if .VisibleFlags}}[global options]{{end}}{{if .Commands}} command [command options]{{end}}
 
 EXAMPLES:
-
 	# Scan a source directory
-	$ {{.HelpName}} scan source -r <source_directory>
+	$ {{.Name}} scan source -r <source_directory>
 
 	# Scan a container image
-	$ {{.HelpName}} scan image <image_name>
+	$ {{.Name}} scan image <image_name>
 
 	# Scan a local image archive (e.g. a tar file) and generate HTML output
-	$ {{.HelpName}} scan image --serve --archive <image_name.tar>
+	$ {{.Name}} scan image --serve --archive <image_name.tar>
 
 	# Fix vulnerabilities in a manifest file and lockfile (non-interactive mode)
-	$ {{.HelpName}} fix --non-interactive -M <manifest_file> -L <lockfile>
+	$ {{.Name}} fix --non-interactive -M <manifest_file> -L <lockfile>
 
-	For full usage details, please refer to the help command of each subcommand (e.g. {{.HelpName}} scan --help).
+	For full usage details, please refer to the help command of each subcommand (e.g. {{.Name}} scan --help).
 
 VERSION:
 	{{.Version}}
